+++ src/app.module.css
@@ -118,24 +118,3 @@ body {
   background: rgb(173, 173, 173); 
 }
 
-/*Icono de audio*/
-.audioOn {
-  position: absolute;
-  right: 0;
-}
-
-.on1 {
-  top: 575px;
-  margin-right: 15px;
-}
-.on2 {
-  top: 575px;
-  right: 55px;
-}
-
-.off{
-  top: 575px;
-  right: 20px;
-  width: 25px;
-  height: 25px;
-}
