+++ src/componentes/footerNav/NavFooter.jsx
@@ -1,6 +1,26 @@
+import React, { useState } from "react";
 import style from "./NavFooter.module.css";
+import on from "../../assets/musicon.png";
+import off from "../../assets/musicoff.png";
+import useSound from "use-sound";
+import musica from "../../assets/musica.mp3";
 
 const Nav = () => {
+  const [play, { stop }] = useSound(musica);
+  const [isPlaying, setIsPlaying] = useState(false);
+  const handlePlay = () => {
+    if (!isPlaying) {
+      play();
+      setIsPlaying(true);
+    }
+  };
+
+  const handleStop = () => {
+    if (isPlaying) {
+      stop();
+      setIsPlaying(false);
+    }
+  };
   // Obtener la hora actual
   const time = new Date();
 
@@ -24,9 +44,29 @@ const Nav = () => {
         />
         <p className={style.start}>Start</p>
       </div>
+      <div className={style.footer}>
+        <img
+          className={`${style.audioOn} ${style.on1}`}
+          src={on}
+          alt=""
+          onClick={handlePlay}
+        />
+        <img
+          className={`${style.audioOn} ${style.off}`}
+          src={off}
+          alt=""
+          onClick={handleStop}
+        />
+        <img
+          className={`${style.audioOn} ${style.on2}`}
+          src={on}
+          alt=""
+          onClick={handlePlay}
+        />
       <p className={style.clock}>
         {hours}:{min} {pmAm}
       </p>
+      </div>
     </div>
   );
 };
